[ISSUE #8242]♻️Extract Proxy Local adapter and managed Broker lifecycle
