+++ rocketmq-proxy-local/Cargo.toml
@@ -1,0 +1,49 @@
+# Copyright 2023 The RocketMQ Rust Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+[package]
+name = "rocketmq-proxy-local"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords = ["rocketmq", "proxy", "local", "broker"]
+categories = ["network-programming", "asynchronous"]
+description = "Embedded Broker-backed local adapter for RocketMQ Proxy"
+rust-version.workspace = true
+
+[features]
+default = []
+observability = ["rocketmq-broker/otel-metrics"]
+tieredstore = ["rocketmq-broker/tieredstore"]
+
+[dependencies]
+rocketmq-broker = { workspace = true }
+rocketmq-error = { workspace = true }
+rocketmq-model = { workspace = true }
+rocketmq-proxy-core = { workspace = true }
+rocketmq-runtime = { workspace = true }
+
+async-trait = "0.1"
+bytes = { workspace = true }
+cheetah-string = { workspace = true }
+serde = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+tracing = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3"
